fix: refactor import statements in ConfigureExplorersStep and NodeSidebar components to avoid circular dependency
